src/requirements.txt
# Dependencias principales
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Providers
openai>=1.0.0
anthropic>=0.25.0
httpx>=0.25.0

# Interfaz de usuario
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Utilidades
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0

# Desarrollo y testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

